+++ R/tw_get_wikipedia_page_qid.R
@@ -5,7 +5,7 @@
 #' @param url A character vector with the full URL to one or more Wikipedia pages. If given, title and language can be left empty.
 #' @param title Title of a Wikipedia page or final parts of its url. If given, url can be left empty, but language must be provided.
 #' @param language Two-letter language code used to define the Wikipedia version to use. Defaults to language set with `tw_set_language()`; if not set, "en". If url given, this can be left empty.
-#' @param action Defaults to "query". Usually either "query" or "parse". In principle, any valid action value, see: https://www.mediawiki.org/w/api.php
+#' @param action Defaults to "query". Usually either "query" or "parse". In principle, any valid action value, see: "https://www.mediawiki.org/w/api.php"
 #'
 #' @return A character vector of base urls to be used with the MediaWiki API
 #' @export